+++ src/Entry/AbstractEntry.php
@@ -10,11 +10,13 @@ abstract class AbstractEntry
 {
     private string $key;
     private ?string $defaultValue;
+    private string $friendlyType;
 
-    public function __construct(string $key, ?string $defaultValue = null)
+    protected function __construct(string $key, ?string $defaultValue, string $friendlyType)
     {
         $this->setKey($key);
         $this->setDefaultValue($defaultValue);
+        $this->friendlyType = $friendlyType;
     }
 
     final public function getKey(): string
@@ -43,8 +45,13 @@ abstract class AbstractEntry
         return $this;
     }
 
+    final public function getFriendlyType(): string
+    {
+        return $this->friendlyType;
+    }
+
     /**
      * @return mixed
      */
-    abstract public function checkValue(string $value);
+    abstract public function checkValue(string $value, bool $default = false);
 }