+++ src/Exception/MissingValueException.php
@@ -11,9 +11,9 @@ class MissingValueException extends RuntimeException
 {
     private AbstractEntry $entry;
 
-    public function __construct(string $message, AbstractEntry $entry)
+    public function __construct(AbstractEntry $entry)
     {
-        parent::__construct($message);
+        parent::__construct('Variable is missing and has no default value.');
 
         $this->entry = $entry;
     }