+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Add function to export values to an simple array
+- Collect all exceptions at buildtime and throw summary exception
+- Friendly type name for all entry types
 
 ## [0.1.0] - 2021-01-05
 ### Added